api/requirements.txt
Django
djangorestframework
django-cors-headers
uwsgi
mysqlclient
psycopg2